fix(model): resolve two latent type/interface bugs ifx rejects
With the build-order hang fixed, ifx's stricter front end surfaces two genuine
standard-conformance bugs that gfortran silently tolerated:

1. error #6633 (mod_PELAGIC_ECOLOGY PELAGIC_KINETICS): dummy AQUABC_CALLED_BEFORE
   was real(DBL) but the actual argument is the integer GLOBAL::CALLED_BEFORE, and
   it is only forwarded to AQUABC_PELAGIC_KINETICS whose dummy is integer. Retyped
   to integer — the whole chain is now consistently integer (also removes a second
   latent real->int mismatch at the forward call). Behaviour-preserving: the value
   is a 0/1 flag, never used arithmetically; the 0D golden regression is unchanged.

2. error #8055 (AQUABC_SEDIMENT_MODEL_1): dummy SED_BURRIAL_RATE_OUTPUTS was
   'optional', which requires every caller to have an explicit interface — the
   external call in mod_SOLVER has none. It is written unconditionally (under
   switch_burial/=0, no present() guard) and the sole caller always passes it, so
   'optional' was both unnecessary and a latent crash risk. Removed.

gfortran build-estas + 0D golden regression: PASS (526x22 match).

+++ SOURCE_CODE/ESTAS/mod_PELAGIC_ECOLOGY.f90
@@ -1368,7 +1368,11 @@ subroutine PELAGIC_KINETICS &
 
     real(kind = DBL), intent(in)    :: TIME
     real(kind = DBL), intent(in)    :: TIME_STEP
-    real(kind = DBL), intent(inout) :: AQUABC_CALLED_BEFORE
+    ! Integer flag (0/1), never used arithmetically here — only forwarded to
+    ! AQUABC_PELAGIC_KINETICS, whose dummy is integer, and it originates from the integer
+    ! GLOBAL::CALLED_BEFORE. Was real(DBL), which ifx rejects as a type mismatch (error #6633)
+    ! against the integer actual argument; gfortran only tolerated it via an implicit interface.
+    integer, intent(inout) :: AQUABC_CALLED_BEFORE
 
     real(kind = DBL), dimension(nkn, nstate)           :: AQUABC_DERIVATIVES
     real(kind = DBL), dimension(nkn, nstate)           :: AQUABC_STATE_VARIABLES
